Added matrix utils
